corrected of observation
