+++ src/component-catalog/form/Form.test.tsx
@@ -1,4 +1,4 @@
-import { render, screen, fireEvent, waitFor } from '@testing-library/react';
+import { render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 
@@ -73,204 +73,104 @@ describe('Form Component', () => {
   });
 
   describe('Field Rendering', () => {
-    it('renders all field types correctly with proper labels and attributes', () => {
+    it('renders all fields', () => {
       render(<Form fields={fields} onSubmit={mockOnSubmit} />);
 
-      expect(screen.getByLabelText('Email')).toBeInTheDocument();
-      expect(screen.getByLabelText('Password')).toBeInTheDocument();
-      expect(screen.getByLabelText('Search with icon')).toBeInTheDocument();
-
-      const selectField = screen.getByLabelText('Role');
-      expect(selectField).toBeInTheDocument();
+      expect(screen.getByLabelText(/email/i)).toBeInTheDocument();
+      expect(screen.getByLabelText(/password/i)).toBeInTheDocument();
+      expect(screen.getByLabelText(/search with icon/i)).toBeInTheDocument();
+      expect(screen.getByLabelText(/role/i)).toBeInTheDocument();
 
       expect(screen.getByText('Gender')).toBeInTheDocument();
       expect(screen.getByLabelText('Male')).toBeInTheDocument();
       expect(screen.getByLabelText('Female')).toBeInTheDocument();
 
-      expect(screen.getByLabelText('Accept Terms')).toBeInTheDocument();
-
-      const textarea = screen.getByLabelText('Bio');
-      expect(textarea).toBeInTheDocument();
-      expect(textarea.tagName).toBe('TEXTAREA');
-      expect(textarea).toHaveAttribute('rows', '4');
-      expect(textarea).toHaveAttribute('placeholder', 'Tell us about yourself...');
+      expect(screen.getByLabelText(/accept terms/i)).toBeInTheDocument();
+      expect(screen.getByLabelText(/bio/i)).toBeInTheDocument();
     });
 
-    it('renders select field with options', () => {
-      render(<Form fields={[fields[3]]} onSubmit={mockOnSubmit} />);
+    it('disables fields correctly', () => {
+      render(<Form fields={fields} onSubmit={mockOnSubmit} disabled />);
 
-      const select = screen.getByLabelText('Role');
-      expect(select).toBeInTheDocument();
-
-      fireEvent.mouseDown(select);
-
-      expect(screen.getByText('Admin')).toBeInTheDocument();
-      expect(screen.getByText('User')).toBeInTheDocument();
-    });
-
-    it('renders radio group with correct options', () => {
-      render(<Form fields={[fields[4]]} onSubmit={mockOnSubmit} />);
-
-      expect(screen.getByText('Gender')).toBeInTheDocument();
-      expect(screen.getByLabelText('Male')).toBeInTheDocument();
-      expect(screen.getByLabelText('Female')).toBeInTheDocument();
-    });
-
-    it('disables fields when disabled prop is true', () => {
-      render(<Form fields={fields} onSubmit={mockOnSubmit} disabled={true} />);
-
-      expect(screen.getByLabelText('Email')).toBeDisabled();
-      expect(screen.getByLabelText('Password')).toBeDisabled();
-      expect(screen.getByLabelText('Search with icon')).toBeDisabled();
-      expect(screen.getByLabelText('Accept Terms')).toBeDisabled();
-      expect(screen.getByLabelText('Bio')).toBeDisabled();
+      expect(screen.getByLabelText(/email/i)).toBeDisabled();
+      expect(screen.getByLabelText(/password/i)).toBeDisabled();
+      expect(screen.getByLabelText(/accept terms/i)).toBeDisabled();
     });
   });
 
-  describe('Validation Errors', () => {
-    it('validates email format correctly', async () => {
+  describe('Validation', () => {
+    it('validates required select', async () => {
       const user = userEvent.setup();
-      render(<Form fields={[fields[0]]} onSubmit={mockOnSubmit} />);
 
-      const emailInput = screen.getByLabelText('Email');
-      await user.type(emailInput, 'invalid-email');
-      await user.tab();
-
-      await waitFor(() => {
-        expect(screen.getByText('Invalid email')).toBeInTheDocument();
-      });
-    });
-
-    it('validates password minimum length', async () => {
-      const user = userEvent.setup();
-      render(<Form fields={[fields[1]]} onSubmit={mockOnSubmit} />);
-
-      const passwordInput = screen.getByLabelText('Password');
-      await user.type(passwordInput, '123');
-      await user.tab();
-
-      await waitFor(() => {
-        expect(screen.getByText('Password must be at least 6 characters')).toBeInTheDocument();
-      });
-    });
-
-    it('validates select field when required', async () => {
-      const user = userEvent.setup();
-      const selectFields = [
+      const selectFields: FormField[] = [
         {
           name: 'role',
           label: 'Role',
-          type: 'select' as const,
+          type: 'select',
           required: true,
-          options: [
-            { value: 'admin', label: 'Admin' },
-            { value: 'user', label: 'User' },
-          ],
+          options: [{ value: 'admin', label: 'Admin' }],
         },
       ];
 
       render(<Form fields={selectFields} onSubmit={mockOnSubmit} />);
 
-      const submitButton = screen.getByText('Submit');
-      await user.click(submitButton);
+      await user.click(screen.getByRole('button', { name: /submit/i }));
 
-      await waitFor(() => {
-        expect(screen.getByText('Role is required')).toBeInTheDocument();
-      });
+      expect(await screen.findByText(/required/i)).toBeInTheDocument();
     });
 
-    it('validates checkbox when required', async () => {
+    it('validates checkbox required', async () => {
       const user = userEvent.setup();
-      const checkboxFields = [
+
+      const checkboxFields: FormField[] = [
         {
           name: 'terms',
           label: 'Accept Terms',
-          type: 'checkbox' as const,
+          type: 'checkbox',
           required: true,
         },
       ];
 
       render(<Form fields={checkboxFields} onSubmit={mockOnSubmit} />);
 
-      const submitButton = screen.getByText('Submit');
-      await user.click(submitButton);
+      await user.click(screen.getByRole('button', { name: /submit/i }));
 
-      await waitFor(() => {
-        expect(screen.getByText('Accept Terms must be accepted')).toBeInTheDocument();
-      });
-    });
-
-    it('clears validation error when field is corrected', async () => {
-      const user = userEvent.setup();
-      render(<Form fields={[fields[0]]} onSubmit={mockOnSubmit} />);
-
-      const emailInput = screen.getByLabelText('Email');
-
-      await user.type(emailInput, 'invalid-email');
-      await user.tab();
-
-      await waitFor(() => {
-        expect(screen.getByText('Invalid email')).toBeInTheDocument();
-      });
-
-      await user.clear(emailInput);
-      await user.type(emailInput, 'valid@example.com');
-      await user.tab();
-
-      await waitFor(() => {
-        expect(screen.queryByText('Invalid email')).not.toBeInTheDocument();
-      });
+      expect(await screen.findByText(/accepted/i)).toBeInTheDocument();
     });
   });
 
-  describe('Form Submit', () => {
-    describe('Form Reset', () => {
-      it('resets radio group correctly', async () => {
-        const user = userEvent.setup();
-        render(<Form fields={[fields[4]]} onSubmit={mockOnSubmit} showResetButton={true} />);
+  describe('Reset', () => {
+    it('resets checkbox', async () => {
+      const user = userEvent.setup();
 
-        await user.click(screen.getByLabelText('Male'));
-        expect(screen.getByLabelText('Male')).toBeChecked();
+      render(<Form fields={[fields[5]]} onSubmit={mockOnSubmit} showResetButton />);
 
-        const resetButton = screen.getByText('Reset');
-        await user.click(resetButton);
+      const checkbox = screen.getByLabelText(/accept terms/i);
 
-        await waitFor(() => {
-          expect(screen.getByLabelText('Male')).not.toBeChecked();
-          expect(screen.getByLabelText('Female')).not.toBeChecked();
-        });
+      await user.click(checkbox);
+      expect(checkbox).toBeChecked();
+
+      await user.click(screen.getByRole('button', { name: /reset/i }));
+
+      await waitFor(() => {
+        expect(checkbox).not.toBeChecked();
       });
+    });
 
-      it('resets checkbox correctly', async () => {
-        const user = userEvent.setup();
-        render(<Form fields={[fields[5]]} onSubmit={mockOnSubmit} showResetButton={true} />);
+    it('resets textarea', async () => {
+      const user = userEvent.setup();
 
-        const checkbox = screen.getByLabelText('Accept Terms');
-        await user.click(checkbox);
-        expect(checkbox).toBeChecked();
+      render(<Form fields={[fields[6]]} onSubmit={mockOnSubmit} showResetButton />);
 
-        const resetButton = screen.getByText('Reset');
-        await user.click(resetButton);
+      const textarea = screen.getByLabelText(/bio/i);
 
-        await waitFor(() => {
-          expect(checkbox).not.toBeChecked();
-        });
-      });
+      await user.type(textarea, 'test');
+      expect(textarea).toHaveValue('test');
 
-      it('resets textarea correctly', async () => {
-        const user = userEvent.setup();
-        render(<Form fields={[fields[6]]} onSubmit={mockOnSubmit} showResetButton={true} />);
+      await user.click(screen.getByRole('button', { name: /reset/i }));
 
-        const textarea = screen.getByLabelText('Bio');
-        await user.type(textarea, 'Some bio text');
-        expect(textarea).toHaveValue('Some bio text');
-
-        const resetButton = screen.getByText('Reset');
-        await user.click(resetButton);
-
-        await waitFor(() => {
-          expect(textarea).toHaveValue('');
-        });
+      await waitFor(() => {
+        expect(textarea).toHaveValue('');
       });
     });
   });
